Change Swig interface size_t to int.

+++ include/cppad/swig/a_fun.hpp
@@ -54,16 +54,16 @@ class CPPAD_SWIG_LIB_PUBLIC a_fun
 		const std::vector<double>& w
 	);
 	// forward
-	std::vector<double> forward(size_t p, const std::vector<double>& xp );
+	std::vector<double> forward(int p, const std::vector<double>& xp );
 	// reverse
-	std::vector<double> reverse(size_t q, const std::vector<double>& yq );
+	std::vector<double> reverse(int q, const std::vector<double>& yq );
 	// optimize
 	void optimize(void);
 	// a_fun properties
-	size_t size_ind(void) const;
-	size_t size_dep(void) const;
-	size_t size_var() const;
-	size_t size_op() const;
+	int size_ind(void) const;
+	int size_dep(void) const;
+	int size_var() const;
+	int size_op() const;
 };
 
 } // END_CPPAD_SWIG_NAMESPACE
